[h264 dec] fixed MSVC 2015 warnings
umc_h264_heap.h(270): warning C4457: declaration of 'obj' hides function parameter
umc_d3d_video_processing.cpp(248): warning C4838: conversion from 'UINT' to 'LONG' requires a narrowing conversion

+++ _studio/shared/umc/codec/color_space_converter/src/umc_d3d_video_processing.cpp
@@ -242,7 +242,9 @@ Status D3DVideoProcessing::GetFrame(MediaData *in, MediaData *out)
     SrcRect.bottom = pInputSurface->offset_y + pInput->GetHeight();
 
     pOutputSurface->pSurface->GetDesc(&SurfaceDescription);
-    ::RECT TrgtRect = {0, 0, SurfaceDescription.Width, SurfaceDescription.Height};
+    // next line we do explicit  UINT -> LONG conversion i.e. sing -> unsign.
+    // it is ok while surface W and H less than 2147483648
+    ::RECT TrgtRect = {0, 0, static_cast<LONG>(SurfaceDescription.Width), static_cast<LONG>(SurfaceDescription.Height)};
 
     //GetClientRect(m_hFocusWindow, &TrgtRect);
     msBlitParams.TargetRect = TrgtRect;
